Cleanup and error condition handling.
