+++ lib/tabs/profile/widgets/username_availability_widget.dart
